requirements.txt
Flask==2.3.3
Authlib==1.4.1
requests==2.31.0
pytz==2023.3
python-dotenv==1.0.0
gunicorn==21.2.20
